Change ComboboxModel to size_t.
Bug: 1292951
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3760854
Commit-Queue: Renato Silva <rrsilva@google.com>
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1024699}

+++ components/autofill/core/browser/ui/country_combobox_model.cc
@@ -76,11 +76,11 @@ void CountryComboboxModel::SetCountries(
             std::back_inserter(countries_));
 }
 
-int CountryComboboxModel::GetItemCount() const {
+size_t CountryComboboxModel::GetItemCount() const {
   return countries_.size();
 }
 
-std::u16string CountryComboboxModel::GetItemAt(int index) const {
+std::u16string CountryComboboxModel::GetItemAt(size_t index) const {
   AutofillCountry* country = countries_[index].get();
   if (country)
     return countries_[index]->name();
@@ -90,7 +90,7 @@ std::u16string CountryComboboxModel::GetItemAt(int index) const {
   return u"---";
 }
 
-bool CountryComboboxModel::IsItemSeparatorAt(int index) const {
+bool CountryComboboxModel::IsItemSeparatorAt(size_t index) const {
   return !countries_[index];
 }
